refactor(#54): moving files around
