+++ packages/pixeloven-core/common/src/Common.ts
@@ -1,9 +1,10 @@
 /**
  * Cleanup and merge options
+ * @todo Should generalize
  * @param defaults
  * @param options
  */
-export function mergeOptions<T>(defaults: T, options: Partial<T>): T {
+function mergeOptions<T>(defaults: T, options: Partial<T>): T {
     Object.keys(options).forEach(key => {
         if (options[key] === undefined) {
             delete options[key];
@@ -13,4 +14,26 @@ export function mergeOptions<T>(defaults: T, options: Partial<T>): T {
         ...defaults,
         ...options,
     };
+}
+
+/**
+ * @todo Move to core
+ */
+type DefinedObjKeys<T> = ({ [P in keyof T]: T[P] extends undefined ? never : P })[keyof T];
+type NonEmptyObject<T, P extends DefinedObjKeys<T> = DefinedObjKeys<T>> = { [PP in P]: T[PP] };
+
+function removeEmpty<T>(input: Array<T | undefined>): T[];
+function removeEmpty<T>(input: { [P in keyof T]: T[P] }): NonEmptyObject<T>;
+function removeEmpty<T>(input: Array<T | undefined> | { [P in keyof T]: T[P] }) {
+    if (Array.isArray(input)) {
+        return input.filter((item) => !!item);
+    }
+    return Object.keys(input)
+        .filter(key => !!input[key])
+        .reduce((res, key) => Object.assign(res, { [key]: input[key] }), {} );
+}
+
+export {
+    mergeOptions,
+    removeEmpty
 }
