+++ packages/pixeloven-webpack/config/package.json
@@ -26,6 +26,8 @@
     "@babel/preset-env": "7.5.5",
     "@babel/preset-react": "7.0.0",
     "@babel/preset-typescript": "7.3.3",
+    "@pixeloven-core/common": "^6.0.0-beta.7",
+    "@pixeloven-core/env": "^6.0.0-beta.7",
     "@pixeloven-core/filesystem": "^6.0.0-beta.7",
     "autoprefixer": "9.6.1",
     "babel-loader": "8.0.6",
